refactor(safe): create intermediary folders
This is so that additional safe logic can be grouped into the safe
folder while keeping the signature verification logic in its own folder.
